chore(aft): Refactor models
Separates the pre-processed and processed aft configurations into `RawAftConfig` and `AftConfig`.
