+++ packages/aft/lib/src/commands/create_command.dart
@@ -66,11 +66,11 @@ class CreateCommand extends AmplifyCommand {
         repoPackages['amplify_core']!.version.amplifyConstraint();
     final amplifyFlutterConstraint =
         repoPackages['amplify_flutter']!.version.amplifyConstraint();
-    final dartSdkConstraint = aftConfig.environment.sdk.toString();
-    final flutterSdkConstraint = aftConfig.environment.flutter.toString();
-    final androidMinSdkVersion = aftConfig.environment.android.minSdkVersion;
-    final iosMinOsVersion = aftConfig.environment.ios.minOSVersion;
-    final macosMinOsVersion = aftConfig.environment.macOS.minOSVersion;
+    final dartSdkConstraint = rawAftConfig.environment.sdk.toString();
+    final flutterSdkConstraint = rawAftConfig.environment.flutter.toString();
+    final androidMinSdkVersion = rawAftConfig.environment.android.minSdkVersion;
+    final iosMinOsVersion = rawAftConfig.environment.ios.minOSVersion;
+    final macosMinOsVersion = rawAftConfig.environment.macOS.minOSVersion;
 
     await generator.hooks.preGen(
       workingDirectory: outputDir.path,
